Move fixtures (WIP).
