+++ tests/test_search_boring.py
@@ -2,29 +2,12 @@
 import datetime
 
 import pytest
-
 from owslib.fes import PropertyIsEqualTo
+
 from pydov.search.boring import BoringSearch
 from pydov.types.boring import Boring
 from pydov.util import owsutil
-from tests.abstract import (
-    AbstractTestSearch,
-)
-
-from tests.test_search import (
-    mp_wfs,
-    wfs,
-    mp_get_schema,
-    mp_remote_md,
-    mp_remote_fc,
-    mp_remote_describefeaturetype,
-    mp_remote_wfs_feature,
-    mp_remote_xsd,
-    mp_dov_xml,
-    mp_dov_xml_broken,
-    wfs_getfeature,
-    wfs_feature,
-)
+from tests.abstract import AbstractTestSearch
 
 location_md_metadata = 'tests/data/types/boring/md_metadata.xml'
 location_fc_featurecatalogue = \
@@ -35,30 +18,6 @@ location_wfs_getfeature = 'tests/data/types/boring/wfsgetfeature.xml'
 location_wfs_feature = 'tests/data/types/boring/feature.xml'
 location_dov_xml = 'tests/data/types/boring/boring.xml'
 location_xsd_base = 'tests/data/types/boring/xsd_*.xml'
-
-
-@pytest.fixture
-def md_metadata(wfs, mp_remote_md):
-    """PyTest fixture providing a MD_Metadata instance of the
-    dov-pub:Boringen layer.
-
-    Parameters
-    ----------
-    wfs : pytest.fixture returning owslib.wfs.WebFeatureService
-        WebFeatureService based on the local GetCapabilities.
-    mp_remote_md : pytest.fixture
-        Monkeypatch the call to get the remote metadata of the
-        dov-pub:Boringen layer.
-
-    Returns
-    -------
-    owslib.iso.MD_Metadata
-        Parsed metadata describing the Boringen WFS layer in more detail,
-        in the ISO 19115/19139 format.
-
-    """
-    contentmetadata = wfs.contents['dov-pub:Boringen']
-    return owsutil.get_remote_metadata(contentmetadata)
 
 
 class TestBoringSearch(AbstractTestSearch):
