+++ tests/test_search_itp_informelehydrogeologischestratigrafie.py
@@ -1,37 +1,16 @@
 """Module grouping tests for the interpretaties search module."""
-import pandas as pd
 import numpy as np
+import pandas as pd
 import pytest
+from owslib.fes import PropertyIsEqualTo
 from pandas import DataFrame
 
 import pydov
-from owslib.fes import PropertyIsEqualTo
 from pydov.search.interpretaties import (
-    FormeleStratigrafieSearch,
-    InformeleHydrogeologischeStratigrafieSearch,
-)
-from pydov.types.interpretaties import (
-    FormeleStratigrafie,
-    InformeleHydrogeologischeStratigrafie,
-)
-from tests.abstract import (
-    AbstractTestSearch,
-)
-
-from tests.test_search import (
-    mp_wfs,
-    wfs,
-    mp_get_schema,
-    mp_remote_md,
-    mp_remote_fc,
-    mp_remote_describefeaturetype,
-    mp_remote_wfs_feature,
-    mp_remote_xsd,
-    mp_dov_xml,
-    mp_dov_xml_broken,
-    wfs_getfeature,
-    wfs_feature,
-)
+    FormeleStratigrafieSearch, InformeleHydrogeologischeStratigrafieSearch)
+from pydov.types.interpretaties import (FormeleStratigrafie,
+                                        InformeleHydrogeologischeStratigrafie)
+from tests.abstract import AbstractTestSearch
 
 location_md_metadata = \
     'tests/data/types/interpretaties/informele_hydrogeologische_stratigrafie' \
