+++ tests/test_search_sondering.py
@@ -3,31 +3,14 @@ import datetime
 
 import pandas as pd
 import pytest
-
 from owslib.fes import PropertyIsEqualTo
+
 from pydov.search.boring import BoringSearch
 from pydov.search.sondering import SonderingSearch
 from pydov.types.boring import Boring
 from pydov.types.sondering import Sondering
 from pydov.util import owsutil
-from tests.abstract import (
-    AbstractTestSearch,
-)
-
-from tests.test_search import (
-    mp_wfs,
-    wfs,
-    mp_get_schema,
-    mp_remote_md,
-    mp_remote_fc,
-    mp_remote_describefeaturetype,
-    mp_remote_wfs_feature,
-    mp_remote_xsd,
-    mp_dov_xml,
-    mp_dov_xml_broken,
-    wfs_getfeature,
-    wfs_feature,
-)
+from tests.abstract import AbstractTestSearch
 
 location_md_metadata = 'tests/data/types/sondering/md_metadata.xml'
 location_fc_featurecatalogue = \
