+++ tests/test_types_itp_formelestratigrafie.py
@@ -3,15 +3,9 @@ pydov.types.interpretaties.FormeleStratigrafie class."""
 from pydov.types.interpretaties import FormeleStratigrafie
 from pydov.util.dovutil import build_dov_url
 from tests.abstract import AbstractTestTypes
-
-from tests.test_search_itp_formelestratigrafie import (
-    wfs_feature,
-    wfs_getfeature,
-    mp_dov_xml,
-    location_wfs_feature,
-    location_wfs_getfeature,
-    location_dov_xml,
-)
+from tests.test_search_itp_formelestratigrafie import (location_dov_xml,
+                                                       location_wfs_feature,
+                                                       location_wfs_getfeature)
 
 
 class TestFormeleStratigrafie(AbstractTestTypes):
